Remove duplication and feature-envy via MethodBasedTestDescriptor
Issue: #14
